Module aws_sdk_proton::model

Modules

account_settings

See AccountSettings

compatible_environment_template

See CompatibleEnvironmentTemplate

compatible_environment_template_input

See CompatibleEnvironmentTemplateInput

environment

See Environment

environment_account_connection

See EnvironmentAccountConnection

environment_account_connection_summary

See EnvironmentAccountConnectionSummary

environment_summary

See EnvironmentSummary

environment_template

See EnvironmentTemplate

environment_template_filter

See EnvironmentTemplateFilter

environment_template_summary

See EnvironmentTemplateSummary

environment_template_version

See EnvironmentTemplateVersion

environment_template_version_summary

See EnvironmentTemplateVersionSummary

s3_object_source

See S3ObjectSource

service

See Service

service_instance

See ServiceInstance

service_instance_summary

See ServiceInstanceSummary

service_pipeline

See ServicePipeline

service_summary

See ServiceSummary

service_template

See ServiceTemplate

service_template_summary

See ServiceTemplateSummary

service_template_version

See ServiceTemplateVersion

service_template_version_summary

See ServiceTemplateVersionSummary

tag

See Tag

Structs

AccountSettings

The AWS Proton pipeline service role data.

CompatibleEnvironmentTemplate

Compatible environment template data.

CompatibleEnvironmentTemplateInput

Compatible environment template data.

Environment

The environment detail data. An AWS Proton environment is a set resources shared across an AWS Proton service.

EnvironmentAccountConnection

The environment account connection detail data.

EnvironmentAccountConnectionSummary

A summary of the environment account connection detail data.

EnvironmentSummary

A summary of the environment detail data.

EnvironmentTemplate

The environment template data.

EnvironmentTemplateFilter

A search filter for environment templates.

EnvironmentTemplateSummary

The environment template data.

EnvironmentTemplateVersion

The environment template version data.

EnvironmentTemplateVersionSummary

A summary of the version of an environment template detail data.

S3ObjectSource

Template bundle S3 bucket data.

Service

The service detail data.

ServiceInstance

The service instance detail data.

ServiceInstanceSummary

A summary of the service instance detail data.

ServicePipeline

The service pipeline detail data.

ServiceSummary

A summary of the service detail data.

ServiceTemplate

The service template detail data.

ServiceTemplateSummary

The service template summary data.

ServiceTemplateVersion

The version of a service template detail data.

ServiceTemplateVersionSummary

A summary of the service template version detail data.

Tag

A description of a resource tag.

Enums

DeploymentStatus

DeploymentUpdateType

EnvironmentAccountConnectionRequesterAccountType

EnvironmentAccountConnectionStatus

Provisioning

ServiceStatus

TemplateVersionSourceInput

TemplateVersionStatus