Module aws_sdk_proton::operation

Structs

AcceptEnvironmentAccountConnection

In a management account, an environment account connection request is accepted. When the environment account connection request is

CancelEnvironmentDeployment

Attempts to cancel an environment deployment on an UpdateEnvironment action, if the deployment is

CancelServiceInstanceDeployment

Attempts to cancel a service instance deployment on an UpdateServiceInstance action, if the deployment is

CancelServicePipelineDeployment

Attempts to cancel a service pipeline deployment on an UpdateServicePipeline action, if the deployment is

CreateEnvironment

Deploy a new environment. An AWS Proton environment is created from an environment template that defines infrastructure and resources

CreateEnvironmentAccountConnection

Create an environment account connection in an environment account so that environment infrastructure resources can be provisioned in

CreateEnvironmentTemplate

Create an environment template for AWS Proton. For more information, see Environment Templates in the AWS Proton Administrator

CreateEnvironmentTemplateVersion

Create a new major or minor version of an environment template. A major version of an environment template is a version that

CreateService

Create an AWS Proton service. An AWS Proton service is an instantiation of a service template and often includes several service instances

CreateServiceTemplate

Create a service template. The administrator creates a service template to define standardized infrastructure and an optional CICD

CreateServiceTemplateVersion

Create a new major or minor version of a service template. A major version of a service template is a version that

DeleteEnvironment

Delete an environment.

DeleteEnvironmentAccountConnection

In an environment account, delete an environment account connection.

DeleteEnvironmentTemplate

If no other major or minor versions of an environment template exist, delete the environment template.

DeleteEnvironmentTemplateVersion

If no other minor versions of an environment template exist, delete a major version of the environment template if it's not the

DeleteService

Delete a service.

DeleteServiceTemplate

If no other major or minor versions of the service template exist, delete the service template.

DeleteServiceTemplateVersion

If no other minor versions of a service template exist, delete a major version of the service template if it's not the

GetAccountSettings

Get detail data for the AWS Proton pipeline service role.

GetEnvironment

Get detail data for an environment.

GetEnvironmentAccountConnection

In an environment account, view the detail data for an environment account connection.

GetEnvironmentTemplate

Get detail data for an environment template.

GetEnvironmentTemplateVersion

View detail data for a major or minor version of an environment template.

GetService

Get detail data for a service.

GetServiceInstance

Get detail data for a service instance. A service instance is an instantiation of service template, which is running in a specific

GetServiceTemplate

Get detail data for a service template.

GetServiceTemplateVersion

View detail data for a major or minor version of a service template.

ListEnvironmentAccountConnections

View a list of environment account connections.

ListEnvironmentTemplateVersions

List major or minor versions of an environment template with detail data.

ListEnvironmentTemplates

List environment templates.

ListEnvironments

List environments with detail data summaries.

ListServiceInstances

List service instances with summaries of detail data.

ListServiceTemplateVersions

List major or minor versions of a service template with detail data.

ListServiceTemplates

List service templates with detail data.

ListServices

List services with summaries of detail data.

ListTagsForResource

List tags for a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administrator Guide or AWS Proton User Guide.

RejectEnvironmentAccountConnection

In a management account, reject an environment account connection from another environment account.

TagResource

Tag a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administrator Guide or AWS Proton User Guide.

UntagResource

Remove a tag from a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administrator Guide or AWS Proton User Guide.

UpdateAccountSettings

Update the AWS Proton pipeline service account settings.

UpdateEnvironment

Update an environment.

UpdateEnvironmentAccountConnection

In an environment account, update an environment account connection to use a new IAM role.

UpdateEnvironmentTemplate

Update an environment template.

UpdateEnvironmentTemplateVersion

Update a major or minor version of an environment template.

UpdateService

Edit a service description or use a spec to add and delete service instances.

UpdateServiceInstance

Update a service instance.

UpdateServicePipeline

Update the service pipeline.

UpdateServiceTemplate

Update a service template.

UpdateServiceTemplateVersion

Update a major or minor version of a service template.